+++ src/main/java/li/cil/bedrockores/common/config/ore/SelectorType.java
@@ -5,6 +5,7 @@ import net.minecraft.world.DimensionType;
 import net.minecraftforge.common.BiomeDictionary;
 import net.minecraftforge.common.BiomeManager;
 
+import javax.annotation.Nullable;
 import java.util.EnumSet;
 import java.util.Set;
 
@@ -12,7 +13,7 @@ public enum SelectorType {
     Type((config, dimension, dimensionType) -> config.dimensionTypes.contains(dimensionType),
          ((config, biome, biomeTypes, biomeDictTypes) -> config.biomeTypes.stream().anyMatch(biomeTypes::contains))),
     Id(((config, dimension, dimensionType) -> config.dimensionIds.contains(dimension)),
-       ((config, biome, biomeTypes, biomeDictTypes) -> config.biomeIds.contains(biome))),
+       ((config, biome, biomeTypes, biomeDictTypes) -> biome == null ? config.biomeIds.isEmpty() : config.biomeIds.contains(biome))),
     Dictionary((config, dimension, dimensionType) -> false, // No dictionary support for world.
                ((config, biome, biomeTypes, biomeDictTypes) -> config.biomeDictTypes.stream().anyMatch(biomeDictTypes::contains)));
 
@@ -33,7 +34,7 @@ public enum SelectorType {
         return worldPredicate.test(config, dimension, dimensionType);
     }
 
-    boolean select(final OreConfigFilter config, final ResourceLocation biome, final EnumSet<BiomeManager.BiomeType> biomeTypes, final Set<BiomeDictionary.Type> biomeDictTypes) {
+    boolean select(final OreConfigFilter config, @Nullable final ResourceLocation biome, final EnumSet<BiomeManager.BiomeType> biomeTypes, final Set<BiomeDictionary.Type> biomeDictTypes) {
         return biomePredicate.test(config, biome, biomeTypes, biomeDictTypes);
     }
 
@@ -46,6 +47,6 @@ public enum SelectorType {
 
     @FunctionalInterface
     private interface BiomePredicate {
-        boolean test(final OreConfigFilter config, final ResourceLocation biome, final EnumSet<BiomeManager.BiomeType> biomeTypes, final Set<BiomeDictionary.Type> biomeDictTypes);
+        boolean test(final OreConfigFilter config, @Nullable final ResourceLocation biome, final EnumSet<BiomeManager.BiomeType> biomeTypes, final Set<BiomeDictionary.Type> biomeDictTypes);
     }
 }